+++ R/batch_geocoding.R
@@ -41,7 +41,7 @@ batch_census <- function(unique_addresses,
   if (verbose == TRUE) display_query(api_url, query_parameters)
   
   # Query API
-  raw_content <- query_api(api_url, query_parameters, mode = 'file', 
+  response <- query_api(api_url, query_parameters, mode = 'file', 
           batch_file = tmp, content_encoding = "ISO-8859-1", timeout = timeout)
   
   # force certain geographies columns to be read in as character instead of numeric
@@ -53,7 +53,7 @@ batch_census <- function(unique_addresses,
          'census_block' = 'character'),
           NA)
   
-  results <- utils::read.csv(text = raw_content, header = FALSE,
+  results <- utils::read.csv(text = response$content, header = FALSE,
        col.names = return_cols,
        colClasses = column_classes,
        fill = TRUE, stringsAsFactors = FALSE,
@@ -110,11 +110,11 @@ verbose = FALSE, api_url = NULL, geocodio_v = 1.6, limit = 1, ...) {
   if (verbose == TRUE) display_query(api_url, query_parameters)
   
   # Query API
-  raw_content <- query_api(api_url, query_parameters, mode = 'list', input_list = address_list, timeout = timeout)
+  response <- query_api(api_url, query_parameters, mode = 'list', input_list = address_list, timeout = timeout)
   
   # Note that flatten here is necessary in order to get rid of the
   # nested dataframes that would cause dplyr::bind_rows (or rbind) to fail
-  content <- jsonlite::fromJSON(raw_content, flatten = TRUE)
+  content <- jsonlite::fromJSON(response$content, flatten = TRUE)
   
   # How results are parsed depends on whether single line addresses or address
   # components were passed
@@ -323,4 +323,109 @@ batch_here <- function(unique_addresses, lat = 'lat', long = 'long', timeout = 2
   if (full_results == FALSE) return(results[c(lat, long)])
   else return(cbind(results[c(lat, long)], results[!names(results) %in% c(lat, long)]))
   
+}
+
+# Batch geocoding with TomTom
+# ... are arguments passed from the geo() function
+# https://developer.tomtom.com/search-api/search-api-documentation-batch-search/asynchronous-batch-submission
+batch_tomtom <- function(unique_addresses, lat = 'lat', long = 'long', 
+                         timeout = 20, full_results = FALSE,
+                         custom_query = list(), verbose = FALSE,
+                         api_url = NULL, limit = 1, ...) {
+  # limit the dataframe to legitimate arguments
+  address_df <- unique_addresses[names(unique_addresses) %in% get_generic_parameters('tomtom', address_only = TRUE)]
+  
+  if (is.null(api_url)) api_url <- 'https://api.tomtom.com/search/2/batch.json'
+  
+  # Construct query - for display only
+  query_parameters <- get_api_query('tomtom',
+                                    list(limit = limit, api_key = get_key('tomtom')),
+                                    custom_parameters = custom_query)
+  
+  if (verbose == TRUE) display_query(api_url, query_parameters)
+  
+  # Some parameters needs to be included on each element
+  # Others (key, etc) should be in the query
+  api_query_params <- query_parameters[names(query_parameters) %in% c('key', 'redirectMode', 'waitTimeSeconds')]
+  q_elements <- query_parameters[!names(query_parameters) %in% c('key', 'redirectMode', 'waitTimeSeconds')]
+  
+  q_string <- ''
+  for (par in seq_len(length(q_elements))) {
+    dlm <- if (par == 1) '?' else '&'
+    q_string <- paste0(q_string, dlm, 
+                       names(q_elements[par]), '=',  q_elements[[par]])
+  }
+  
+  # Construct body
+  address_list <- list(batchItems = list())
+  
+  for (index in 1:nrow(address_df)) {
+    address_list$batchItems[[index]] <- list(query = paste0('/geocode/', as.list(address_df[index, ]), '.json', q_string))
+  }
+  
+  
+  # Query API
+  response <- httr::POST(api_url, query = api_query_params, 
+                         body = as.list(address_list), 
+                         encode = 'json', httr::timeout(60 * timeout))
+  
+  httr::warn_for_status(response)
+  
+  # https://developer.tomtom.com/search-api/search-api-documentation-batch-search/asynchronous-batch-submission#response-data
+  # if status code is not 200 we have to perform a GET and download the batch asynchronously
+  # On 200 the batch is provided in the response object
+  if (httr::status_code(response) != '200') {
+    if (verbose) message('Asynchronous Batch Download')
+    
+    # A HTTP Response with a Location header that points where the batch results can be obtained.
+    location <- httr::headers(response)$location
+    
+    status <- httr::status_code(response)
+    while (status %in% c('202', '303')) {
+      Sys.sleep(2) # Arbitrary
+      batch_response <- httr::GET(paste0('https://api.tomtom.com', location))
+      status <- httr::status_code(batch_response)
+      if (verbose) httr::message_for_status(batch_response)
+    }
+    
+    if (status == '200') {
+      if (verbose) message("Batch downloaded")
+      raw_content <- httr::content(batch_response, as = 'text', encoding = 'UTF-8')
+    } else {
+      stop(paste0('Batch failed. Status code: ', status), call. = TRUE)
+    }
+    
+  } else {
+    raw_content <- httr::content(response, as = 'text', encoding = 'UTF-8')
+  }
+  
+  # Note that flatten here is necessary in order to get rid of the
+  # nested dataframes that would cause dplyr::bind_rows (or rbind) to fail
+  content <- jsonlite::fromJSON(raw_content)
+  
+  # result_list is a list of dataframes
+  result_list <- content$batchItems$response$results
+  
+  # if no results are returned for a given address then there is a 0 row dataframe in this
+  # list and we need to replace it with a 1 row NA dataframe to preserve the number of rows
+  result_list_filled <- lapply(result_list, filler_df, c('lat', 'long'))
+  
+  # combine list of dataframes into a single tibble. Column names may differ between the dataframes
+  results <- dplyr::bind_rows(result_list_filled)
+  
+  # rename lat/long columns
+  results$lat <- results$position$lat
+  results$long <- results$position$lon
+  
+  # extract address column dataframe
+  tomtom_address <- results$address
+  names(tomtom_address) <- paste0('tomtom_address.', names(tomtom_address))
+  results <- results[!names(results) %in% c('address')]
+  results <- tibble::as_tibble(cbind(results, tomtom_address))
+  
+  names(results)[names(results) == 'lat'] <- lat
+  names(results)[names(results) == 'long'] <- long
+  
+  if (full_results == FALSE) return(results[c(lat, long)])
+  else return(cbind(results[c(lat, long)], results[!names(results) %in% c(lat, long)]))
 }