+++ R/data.R
@@ -56,5 +56,6 @@
 #'  \item \href{https://opencagedata.com/api}{OpenCage}
 #'  \item \href{https://docs.mapbox.com/api/search/geocoding/}{Mapbox}
 #'  \item \href{https://developer.here.com/documentation/geocoding-search-api/dev_guide/index.html}{HERE}
+#'  \item \href{https://developer.tomtom.com/search-api/search-api-documentation-geocoding/geocode}{TomTom}
 #' }     
 "api_parameter_reference"