+++ R/utils.R
@@ -39,7 +39,8 @@ extract_results <- function(method, response, full_results = TRUE, flatten = TRU
       'lat' = response$features$center[[1]][2],
       'long' = response$features$center[[1]][1]
     ), # mapbox results are nested unnames lists
-    'here' = response$items$position[c('lat','lng')]
+    'here' = response$items$position[c('lat','lng')],
+    'tomtom' = response$results$position[c('lat', 'lon')]
   )
   
   # if null result then return NA
@@ -62,9 +63,10 @@ extract_results <- function(method, response, full_results = TRUE, flatten = TRU
       'google' = response$results,
       'opencage' = response$results[!names(response$results) %in% c('geometry')],
       'mapbox' = response$features,
-      'here' = response$items
+      'here' = response$items,
+      'tomtom' = response$results
     )
-    
+
     # add prefix to variable names that likely could be in our input dataset
     # to avoid variable name overlap
     for (var in c('address')) {
@@ -111,7 +113,8 @@ extract_reverse_results <- function(method, response, full_results = TRUE, flatt
                     'google' = response$results[1, ]['formatted_address'],
                     'opencage' = response$results['formatted'],
                     'mapbox' = response$features['place_name'],
-                    'here' = response$items['title']
+                    'here' = response$items['title'],
+                    'tomtom' = response$addresses$address['freeformAddress']
   )
   
   # extract other results (besides single line address)
@@ -126,7 +129,8 @@ extract_reverse_results <- function(method, response, full_results = TRUE, flatt
                       'google' = response$results[1, ][!names(response$results) %in% c('formatted_address')], 
                       'opencage' = response$results[!names(response$results) %in% c('formatted')],
                       'mapbox' = response$features[!names(response$features) %in% c('place_name')],
-                      'here' = response$items[!names(response$items) %in% c('title')]
+                      'here' = response$items[!names(response$items) %in% c('title')],
+                      'tomtom' = response$addresses
     )
     
     # add prefix to variable names that likely could be in our input dataset
@@ -147,52 +151,57 @@ extract_reverse_results <- function(method, response, full_results = TRUE, flatt
   else return(combined_results)
 }
 
-# Check the results of the geocoder service. Display an error message if there is one.
-# returns TRUE if there are problems (errors or blank results) and
-# FALSE if there are no problems detected
-check_results_for_problems <- function(method, raw_results, verbose) {
-
-  # if results are blank
-  if (length(raw_results) == 0) {
-    if (verbose == TRUE) message("No results found")
-  }
-  else if ((method == 'osm') & ("error" %in% names(raw_results))) {
-    message(paste0('Error: ', raw_results$error$message))
-  }
-  else if ((method == 'iq') & ("error" %in% names(raw_results))) {
-    message(paste0('Error: ', raw_results$error))
-  }
-  else if ((method == 'mapbox') & (!is.data.frame(raw_results$features))) {
-    if ("message" %in% names(raw_results)) {
-      message(paste0('Error: ', raw_results$message))
-    }
-  }
-  else if ((method == 'census') & ('errors' %in% names(raw_results))) {
-    message(paste0('Error: ', raw_results$errors))
-  }
-  else if ((method == 'opencage') & (!is.data.frame(raw_results$results))) {
-    if (!is.null(raw_results$status$message)) {
-    message(paste0('Error: ', raw_results$status$message))
-    }
-  }
-  else if ((method == 'geocodio') & (!is.data.frame(raw_results$results))) {
-    if ("error" %in% names(raw_results)) {
-     message(paste0('Error: ', raw_results$error))
-    }
-  }
-  else if ((method == 'google') & (!is.data.frame(raw_results$results))) {
-    if ("error_message" %in% names(raw_results)) {
-      message(paste0('Error: ', raw_results$error_message))
-    }
-  }
-  else if ((method == 'here') & (!is.data.frame(raw_results$items))) {
-    if ("error_description" %in% names(raw_results)) message(paste0('Error: ', raw_results$error_description))
-    else if ("title" %in% names(raw_results)) message(paste0('Error: ', raw_results$title))
+# Extracts errors from a raw response and displays them
+extract_errors_from_results <- function(method, response, verbose) {
+  # test if response contains JSON content
+  if (!jsonlite::validate(response)) {
+    # tomtom does not return JSON content on errors 
+    # in cases like this, display the raw content but limit the length
+    # in case it is really long.
+    message(paste0('Error: ', strtrim(as.character(response), 100)))
   }
   else {
-    return(FALSE) # no problems
+    # parse JSON content
+    raw_results <- jsonlite::fromJSON(response)
+    
+    # if results are blank
+    if (length(raw_results) == 0) {
+      if (verbose == TRUE) message("No results found")
+    }
+    else if ((method == 'osm') & ("error" %in% names(raw_results))) {
+      message(paste0('Error: ', raw_results$error$message))
+    }
+    else if ((method == 'iq') & ("error" %in% names(raw_results))) {
+      message(paste0('Error: ', raw_results$error))
+    }
+    else if ((method == 'mapbox') & (!is.data.frame(raw_results$features))) {
+      if ("message" %in% names(raw_results)) {
+        message(paste0('Error: ', raw_results$message))
+      }
+    }
+    else if ((method == 'census') & ('errors' %in% names(raw_results))) {
+      message(paste0('Error: ', raw_results$errors))
+    }
+    else if ((method == 'opencage') & (!is.data.frame(raw_results$results))) {
+      if (!is.null(raw_results$status$message)) {
+      message(paste0('Error: ', raw_results$status$message))
+      }
+    }
+    else if ((method == 'geocodio') & (!is.data.frame(raw_results$results))) {
+      if ("error" %in% names(raw_results)) {
+       message(paste0('Error: ', raw_results$error))
+      }
+    }
+    else if ((method == 'google') & (!is.data.frame(raw_results$results))) {
+      if ("error_message" %in% names(raw_results)) {
+        message(paste0('Error: ', raw_results$error_message))
+      }
+    }
+    else if ((method == 'here') & (!is.data.frame(raw_results$items))) {
+      if ("error_description" %in% names(raw_results)) message(paste0('Error: ', raw_results$error_description))
+      else if ("title" %in% names(raw_results)) message(paste0('Error: ', raw_results$title))
+    } 
   }
-  return(TRUE) # there was an error or results were blank
 }
 
 # For a list of dataframes, creates an NA df with 1 row with the column name supplied