+++ tests/testthat/test-utils.R
@@ -69,6 +69,8 @@ test_that("Test API Query Creation Functions", {
   
   # loop through all methods and produce queries
   for (method in unique(tidygeocoder::api_parameter_reference[['method']])) {
+    # label to include in error message so we know which method failed
+    method_label = paste0('method = "', method, '"', ' ')
     
     # test overlap between generic and custom parameters
     expect_warning(tidygeocoder::get_api_query(method,
@@ -92,8 +94,8 @@ test_that("Test API Query Creation Functions", {
     # adding to the default_q list
     expect_mapequal(custom_q, c(default_q, cust_arg_list))
     
-    # mapbox address is removed from parameters and put into API URL so this test doesn't apply
-    if (method != 'mapbox') {
+    # mapbox/tomtom address is removed from parameters and put into API URL so this test doesn't apply
+    if (!method %in% c('mapbox', 'tomtom')) {
       expect_mapequal(address_q, c(default_q, 
         tidygeocoder:::create_api_parameter(method, 'address', address_val)))
     }